Add default stat label support to StatCollector
Prometheus supports labelled stats. These may share a metric name but
should have a unique set of name and labels.

As Prometheus will ingest stats from multiple buckets, all existing
stats will be labelled with the bucket name. Similar labelling will be
present for scopes and collections.

As these labels are common to groups of stats, and are not tied to a
_specific_ stat, they are tracked by the StatCollector, rather than
being specified on a per-stat basis.

Proper support for per-stat labels to allow grouping of stats under
metric families will be added in a later patch.

Change-Id: I3ef48edfe2d3c8ca9e0ab5a4d25457aca5ff074f