Introduce units for statistics
While many existing stats represent a value with units (e.g.,
microseconds), this was not always programatically determinable.

In order to correctly expose stats for Prometheus, stats need to be
scaled to base units (microseconds->seconds, kilobytes->bytes) and
appropriately suffixed ( https://prometheus.io/docs/practices/naming/
).

This patch introduces Unit types through which values can be normalised
and the suffix determined.

Units are not yet used in this patch, but will in future be used by
static statistic definitions. They will be constructed once, ahead of
time and used every time a prometheus stat needs to be formatted.

Change-Id: I656f13406e8b039e54c84bdb464c45012ee92c1f