Introduce X-macro statistics definitions
With the introduction of Prometheus as a stats output, more information
is needed for each statistic. Rather than inline everything, stats can
now be declared in stats.def.h and referenced through the
cb::stats::StatKey enum.

Stats can be declared with:

 - units (e.g., microseconds)
 - metric family name
 - key-value labels

These new pieces of information are not used by the existing CBStats
output (but could be in the future). They are added to support
Prometheus stats. Before exposition, stats must be scaled to base units
(seconds, bytes - see https://prometheus.io/docs/practices/naming/ )
and appropriately suffixed.

Metric family names may be used to logically group stats which
represent variants of the same underlying stat. Rather than having
stats like "get_cmd" and "set_cmd" which are only identifiable as
related by the suffix convention, Prometheus can instead represent
these as two stats with the same metric name but different labels
(for example)

 cmd_time_taken_seconds{"op"="get"}
 cmd_time_taken_seconds{"op"="set"}

To do so, the stats could be declared as:

 STAT(get_cmd, microseconds, "cmd_time_taken", "op", "get")
 STAT(set_cmd, microseconds, "cmd_time_taken", "op", "set")

Note, the unit indicates the unit of the value tracked in kv_engine -
Prometheus exposition will _always_ scale the stat to the base unit
from whatever unit is provided - here the value would be divided by
10e6 to get a quantity in seconds.

Given the large number of existing stats, this functionality is all
opt-in for now. CBStats output will remain unchanged.

Change-Id: I3789681c1a35082a34c9ca90992a257f24a02f28